+++ whisper/Extensions/FirstKeyWindow.swift
@@ -1,0 +1,15 @@
+// Copyright 2024 Daniel C Brotsky.  All rights reserved.
+//
+// All material in this project and repository is licensed under the
+// GNU Affero General Public License v3. See the LICENSE file for details.
+
+import SwiftUI
+
+extension UIApplication {
+	var firstKeyWindow: UIWindow? {
+		return UIApplication.shared.connectedScenes
+			.compactMap { $0 as? UIWindowScene }
+			.filter { $0.activationState == .foregroundActive }
+			.first?.keyWindow
+	}
+}
