+++ whisper/Views/MainView/ChoiceView.swift
@@ -280,15 +280,6 @@ struct ChoiceView: View {
     }
 }
 
-extension UIApplication {
-    var firstKeyWindow: UIWindow? {
-        return UIApplication.shared.connectedScenes
-            .compactMap { $0 as? UIWindowScene }
-            .filter { $0.activationState == .foregroundActive }
-            .first?.keyWindow
-    }
-}
-
 #Preview {
     ChoiceView(mode: makeBinding(.ask),
                conversation: makeBinding(nil),
